khuthon ai model/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
omegaconf